refactor: renamed and refactored files
