+++ packages/contracts/test/unit-testing/plugin.ts
@@ -2,8 +2,8 @@ import {PLUGIN_CONTRACT_NAME} from '../../plugin-settings';
 import {DAO, MyPlugin, MyPlugin__factory} from '../../typechain';
 import '../../typechain/src/MyPlugin';
 import {deployWithProxy} from '../../utils/helpers';
+import {STORE_PERMISSION_ID} from '../helpers/constants';
 import {deployTestDao} from '../helpers/test-dao';
-import {STORE_PERMISSION_ID} from './simple-storage-common';
 import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 import {expect} from 'chai';
 import {BigNumber} from 'ethers';
